+++ gui/image/transparent/image_view_for_image_transparent.cpp
@@ -30,7 +30,8 @@ void ImageViewForImageTransparent::onLabelClicked(const QPoint &point)
         const qreal xOffset = (ui_image()->width() - ui_image()->pixmap().width()) / 2.0;
         const qreal yOffset = (ui_image()->height() - ui_image()->pixmap().height()) / 2.0;
 
-        const auto pixel = QPoint((point.x() - xOffset) / scale, (point.y() - yOffset) / scale);
+        const auto pixel = QPoint(static_cast<int>((point.x() - xOffset) / scale),
+                                  static_cast<int>((point.y() - yOffset) / scale));
 
         if (0 <= pixel.x() && 0 <= pixel.y() && pixel.x() < original.width() &&
             pixel.y() < original.height()) {
